Update script installation method
In https://github.com/cisagov/ansible-role-cyhy-core/pull/28 I updated the
installation method for the cisagov/cyhy-core package to directly install with
pip instead of downloading the tarball, extracting, and then installing. This
inadvertently broke this role, and with the decision to switch CyHy environment
redeploys to quarterly (from monthly), this was not noticed until now.

This role will now directly download the desired script to the desired
location. I have left the cisagov/ansible-role-cyhy-core dependency in
place because the downloaded script relies on the functionality of the
cisagov/cyhy-core package.