Specify a version of cisagov/cyhy-core to use
Instead of pulling the `cyhy_archive.sh` script from the head of the
`develop` branch we will instead pull from the git reference specified
in the new `cyhy_archive_cyhy_core_version` variable. This has a
default value of `v1.1.2` (the latest version tag at this time). The
README is updated to include documentation for this new variable as
well.